+++ src/ayame/ayame_websocket_client.h
@@ -39,6 +39,8 @@ class AyameWebsocketClient
   WatchDog watchdog_;
 
   bool connected_;
+  bool is_send_offer_;
+  bool has_is_exist_user_flag_;
 
   webrtc::PeerConnectionInterface::IceServers ice_servers_;
 
@@ -79,6 +81,7 @@ class AyameWebsocketClient
   void onHandshake(boost::system::error_code ec);
   void doRegister();
   void doSendPong();
+  void setIceServersFromConfig(nlohmann::json json_message);
   void createPeerConnection();
 
  public:
@@ -110,6 +113,7 @@ class AyameWebsocketClient
  private:
   void doIceConnectionStateChange(
       webrtc::PeerConnectionInterface::IceConnectionState new_state);
+  void doSetDescription(webrtc::SdpType type);
 };
 
 #endif  // AYAME_WEBSOCKET_CLIENT_