+++ src/main.cpp
@@ -27,6 +27,10 @@
 #endif
 #endif
 
+#if USE_SDL2
+#include "sdl_renderer/sdl_renderer.h"
+#endif
+
 #include "ayame/ayame_server.h"
 #include "connection_settings.h"
 #include "p2p/p2p_server.h"
@@ -93,8 +97,19 @@ int main(int argc, char* argv[]) {
   }
 #endif
 
+#if USE_SDL2
+  std::unique_ptr<SDLRenderer> sdl_renderer = nullptr;
+  if (cs.use_sdl) {
+    sdl_renderer.reset(
+        new SDLRenderer(cs.window_width, cs.window_height, cs.fullscreen));
+  }
+
   std::unique_ptr<RTCManager> rtc_manager(
-      new RTCManager(cs, std::move(capturer)));
+      new RTCManager(cs, std::move(capturer), sdl_renderer.get()));
+#else
+  std::unique_ptr<RTCManager> rtc_manager(
+      new RTCManager(cs, std::move(capturer), nullptr));
+#endif
 
   {
     boost::asio::io_context ioc{1};
@@ -128,9 +143,29 @@ int main(int argc, char* argv[]) {
           ->run();
     }
 
+#if USE_SDL2
+    if (sdl_renderer) {
+      sdl_renderer->SetDispatchFunction(
+        [&ioc](std::function<void ()> f) {
+          if (ioc.stopped()) return;
+          boost::asio::dispatch(ioc.get_executor(), f);
+        });
+
+      ioc.run();
+
+      sdl_renderer->SetDispatchFunction(nullptr);
+    } else {
+      ioc.run();
+    }
+#else
     ioc.run();
+#endif
   }
 
+  //この順番は綺麗に落ちるけど、あまり安全ではない
+#if USE_SDL2
+  sdl_renderer = nullptr;
+#endif
   rtc_manager = nullptr;
 
   return 0;