+++ src/rtc/manager.cpp
@@ -36,11 +36,17 @@
 #include "api/video_codecs/video_encoder_factory.h"
 #include "hw_video_encoder_factory.h"
 #endif
+#if USE_JETSON_ENCODER
+#include "api/video_codecs/video_decoder_factory.h"
+#include "hw_video_decoder_factory.h"
+#endif
 
 RTCManager::RTCManager(
     ConnectionSettings conn_settings,
-    rtc::scoped_refptr<ScalableVideoTrackSource> video_track_source)
-    : _conn_settings(conn_settings) {
+    rtc::scoped_refptr<ScalableVideoTrackSource> video_track_source,
+    VideoTrackReceiver* receiver)
+    : _conn_settings(conn_settings),
+      _receiver(receiver) {
   rtc::InitializeSSL();
 
   _networkThread = rtc::Thread::CreateWithSocketServer();
@@ -97,8 +103,14 @@ RTCManager::RTCManager(
   media_dependencies.video_encoder_factory =
       webrtc::CreateBuiltinVideoEncoderFactory();
 #endif
+#if USE_JETSON_ENCODER
+  media_dependencies.video_decoder_factory =
+      std::unique_ptr<webrtc::VideoDecoderFactory>(
+          absl::make_unique<HWVideoDecoderFactory>());
+#else
   media_dependencies.video_decoder_factory =
       webrtc::CreateBuiltinVideoDecoderFactory();
+#endif
 #endif
   media_dependencies.audio_mixer = nullptr;
   media_dependencies.audio_processing =
@@ -152,6 +164,9 @@ RTCManager::RTCManager(
         _video_track->set_content_hint(
             webrtc::VideoTrackInterface::ContentHint::kText);
       }
+      if (_receiver != nullptr && _conn_settings.show_me) {
+        _receiver->AddTrack(_video_track);
+      }
     } else {
       RTC_LOG(LS_WARNING) << __FUNCTION__ << ": Cannot create video_track";
     }
@@ -175,7 +190,7 @@ std::shared_ptr<RTCConnection> RTCManager::createConnection(
   rtc_config.enable_dtls_srtp = true;
   rtc_config.sdp_semantics = webrtc::SdpSemantics::kUnifiedPlan;
   std::unique_ptr<PeerConnectionObserver> observer(
-      new PeerConnectionObserver(sender));
+      new PeerConnectionObserver(sender, _receiver));
   rtc::scoped_refptr<webrtc::PeerConnectionInterface> connection =
       _factory->CreatePeerConnection(rtc_config, nullptr, nullptr,
                                      observer.get());