+++ src/rtc/observer.h
@@ -4,30 +4,38 @@
 #include "api/peer_connection_interface.h"
 
 #include "messagesender.h"
+#include "video_track_receiver.h"
 
 class PeerConnectionObserver : public webrtc::PeerConnectionObserver {
  public:
-  PeerConnectionObserver(RTCMessageSender* sender) : _sender(sender){};
+  PeerConnectionObserver(RTCMessageSender* sender,
+                         VideoTrackReceiver *receiver)
+                         : _sender(sender),
+                           _receiver(receiver) {};
+  ~PeerConnectionObserver();
 
  protected:
   void OnSignalingChange(
       webrtc::PeerConnectionInterface::SignalingState new_state) override {}
-  void OnAddStream(
-      rtc::scoped_refptr<webrtc::MediaStreamInterface> stream) override {}
-  void OnRemoveStream(
-      rtc::scoped_refptr<webrtc::MediaStreamInterface> stream) override {}
   void OnDataChannel(
       rtc::scoped_refptr<webrtc::DataChannelInterface> data_channel) override {}
   void OnRenegotiationNeeded() override {}
-  void OnIceConnectionChange(
+  void OnStandardizedIceConnectionChange(
       webrtc::PeerConnectionInterface::IceConnectionState new_state) override;
   void OnIceGatheringChange(
       webrtc::PeerConnectionInterface::IceGatheringState new_state) override{};
   void OnIceCandidate(const webrtc::IceCandidateInterface* candidate) override;
-  void OnIceConnectionReceivingChange(bool receiving) override {}
+  void OnTrack(
+      rtc::scoped_refptr<webrtc::RtpTransceiverInterface> transceiver) override;
+  void OnRemoveTrack(
+      rtc::scoped_refptr<webrtc::RtpReceiverInterface> receiver) override;
 
  private:
+  void ClearAllRegisteredTracks();
+
   RTCMessageSender* _sender;
+  VideoTrackReceiver *_receiver;
+  std::vector<webrtc::VideoTrackInterface*> _video_tracks;
 };
 
 class CreateSessionDescriptionObserver